nightly-build-2.5.2
