--- octadist/__init__.py
@@ -1,16 +1,0 @@
-from octadist.main import program_version, program_revision
-
-name = "octadist"
-
-__all__ = [
-    "calc",
-    "coord",
-    "draw",
-    "elements",
-    "linear",
-    "main",
-    "plane",
-    "popup",
-    "projection",
-    "tools",
-]
