+++ octadist_gui/src/coord.py
@@ -1,0 +1,492 @@
+# OctaDist  Copyright (C) 2019  Rangsiman Ketkaew et al.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+# ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+
+from operator import itemgetter
+
+import numpy as np
+
+from octadist_gui.src import echo_logs, echo_outs, elements, linear, popup
+
+
+def count_line(file):
+    """Count line in input file
+
+    :param file: absolute path of input file
+    :type file: str
+    :return: number of line in file
+    :rtype: int
+    """
+    with open(file) as f:
+        for i, l in enumerate(f):
+            pass
+
+    return i + 1
+
+
+def get_coord(self, f):
+    """Check file type, read data, extract atom and coord from input file
+
+    **Support file type**::
+        - XYZ
+        - Gaussian
+        - NWChem
+        - ORCA
+        - Q-Chem
+
+    :param self: master frame
+    :param f: user input filename
+    :type f: str
+    :return a_full: full atomic labels of complex
+    :return c_full: full atomic coordinates of complex
+    :rtype a_full: list
+    :rtype c_full: array
+    """
+    a_full = []
+    c_full = []
+    check = True
+
+    # Check file extension
+    if f.endswith(".xyz"):
+        if check_xyz_file(self, f):
+            echo_logs(self, "Info: Check file type: XYZ file")
+            a_full, c_full = get_coord_xyz(f)
+        else:
+            ftype = "XYZ"
+            popup.err_invalid_ftype(self, ftype)
+            echo_logs(self, "Error: Could not read data in XYZ file {0}".format(f))
+            check = False
+
+    elif f.endswith(".out") or f.endswith(".log"):
+        if check_gaussian_file(f):
+            echo_logs(self, "Info: Check file type: Gaussian Output")
+            a_full, c_full = get_coord_gaussian(f)
+        elif check_nwchem_file(f):
+            echo_logs(self, "Info: Check file type: NWChem Output")
+            a_full, c_full = get_coord_nwchem(f)
+        elif check_orca_file(f):
+            echo_logs(self, "Info: Check file type: ORCA Output")
+            a_full, c_full = get_coord_orca(f)
+        elif check_qchem_file(f):
+            echo_logs(self, "Info: Check file type: Q-Chem Output")
+            a_full, c_full = get_coord_qchem(f)
+        else:
+            echo_logs(self, "Error: Could not read output file {0}".format(f))
+            check = False
+
+    else:
+        text1 = "Error: Could not read file {0}".format(f)
+        echo_outs(self, text1)
+        echo_logs(self, text1)
+        text2 = "Error: File type is not supported by the current version of OctaDist\n"
+        echo_outs(self, text2)
+        echo_logs(self, text2)
+
+        popup.err_wrong_format(self)
+        check = False
+
+    # return values
+    if check:
+        # Remove empty string in list
+        a_full = list(filter(None, a_full))
+        return a_full, c_full
+
+    else:
+        a_full = []
+        c_full = []
+        return a_full, c_full
+
+
+def count_metal(self, a_full, c_full):
+    """Count the number of metal center atom in complex
+
+    :param self: master frame
+    :param a_full: full atomic labels of complex
+    :param c_full: full atomic coordinates of complex
+    :type a_full: list
+    :type c_full: list, array
+    :return count: the total number of metal center atom
+    :return a_metal: atomic labels of metal center atom
+    :return c_metal: atomic coordinates of metal center atom
+    :rtype count: int
+    :rtype a_metal: list
+    :rtype c_metal: array
+    """
+    count = 0
+    a_metal = []
+    c_metal = []
+
+    for i in range(len(a_full)):
+        number = elements.check_atom(a_full[i])
+        if 21 <= number <= 30 or 39 <= number <= 48 or 57 <= number <= 80 or 89 <= number <= 109:
+            count += 1
+            a_metal.append(a_full[i])
+            c_metal.append(c_full[i])
+
+    echo_logs(self, "Info: The complex has {0} metal center atom".format(count))
+    echo_logs(self, "")
+    echo_logs(self, "Info: Show the coordinates of transition metal found in complex")
+    echo_logs(self, "")
+    echo_logs(self, "      Atom        X             Y             Z")
+    echo_logs(self, "      ----    ----------    ----------    ----------")
+    for i in range(count):
+        echo_logs(self, "       {0:>2}   {1:12.8f}  {2:12.8f}  {3:12.8f}"
+                  .format(a_metal[i], c_metal[i][0], c_metal[i][1], c_metal[i][2]))
+    echo_logs(self, "")
+
+    return count, a_metal, c_metal
+
+
+def search_octa(a_full, c_full, c_metal):
+    """Search the octahedral structure in complex
+
+    :param a_full: full atomic labels of complex
+    :param c_full: full atomic coordinates of complex
+    :param c_metal: atomic coordinates of metal center atom
+    :type a_full: list
+    :type c_full: list, array
+    :type c_metal: list, array
+    :return a_octa: atomic labels of octahedral structure
+    :return c_octa: atomic coordinates of octahedral structure
+    :rtype a_octa: list
+    :rtype c_octa: array
+    """
+    dist_list = []
+    for i in range(len(a_full)):
+        dist = linear.distance_bwn_points(c_metal, c_full[i])
+        dist_list.append([a_full[i], c_full[i], dist])
+
+    dist_list.sort(key=itemgetter(2))  # sort list of tuples by distance in ascending order
+    dist_list = dist_list[:7]  # Get only first 7 atoms
+    a_octa, c_octa, distance = zip(*dist_list)  # Collect atom and coordinates
+
+    # list --> array
+    c_octa = np.asarray(c_octa)
+
+    return a_octa, c_octa
+
+
+def check_xyz_file(self, f):
+    """Check if the input file is .xyz file format
+
+    xyz file format
+    ---------------
+
+    <number of atom>
+    comment
+    <index 0> <X> <Y> <Z>
+    <index 1> <X> <Y> <Z>
+    ...
+    <index 6> <X> <Y> <Z>
+
+    :param self: master frame
+    :param f: user input filename
+    :type f: str
+    :return: if the file is .xyz format, return True
+    :rtype: bool
+
+    **Example**
+        7
+        Comment: From Excel file
+        Fe  6.251705    9.063211    5.914842
+        N   8.15961     9.066456    5.463087
+        N   6.749414    10.457551   7.179682
+        N   5.709997    10.492955   4.658257
+        N   4.350474    9.106286    6.356091
+        O   5.789096    7.796326    4.611355
+        O   6.686381    7.763872    7.209699
+
+    """
+    # import extract module here to avoid circular dependencies
+
+    file = open(f, "r")
+
+    first_line = file.readline()
+
+    # Check if the first line is integer
+    try:
+        int(first_line)
+    except ValueError:
+        echo_logs(self, "Error: The first line of XYZ file must be the total number of atoms")
+        return False
+
+    if count_line(f) < 9:
+        echo_logs(self, "Error: The complex must contain at least 1 metal center and 6 ligand atoms.")
+        return False
+    else:
+        return True
+
+
+def get_coord_xyz(f):
+    """Get coordinate from .xyz file
+
+    :param f: user input filename
+    :type f: str
+    :return a_full: full atomic labels of complex
+    :return c_full: full atomic coordinates of complex
+    :rtype a_full: list
+    :rtype c_full: array
+    """
+    file = open(f, "r")
+    # read file from 3rd line
+    line = file.readlines()[2:]
+    file.close()
+
+    a_full = []
+    for l in line:
+        # read atom on 1st column and insert to list
+        l_strip = l.strip()
+        lst = l_strip.split(' ')[0]
+        a_full.append(lst)
+
+    file = open(f, "r")
+    c_full = np.loadtxt(file, skiprows=2, usecols=[1, 2, 3])
+    file.close()
+
+    return a_full, c_full
+
+
+def check_gaussian_file(f):
+    """Check if the input file is Gaussian file format
+
+    :param f: user input filename
+    :type f: str
+    :return: if file is Gaussian output file, return True
+    :rtype: bool
+    """
+    gaussian_file = open(f, "r")
+    nline = gaussian_file.readlines()
+    for i in range(len(nline)):
+        if "Standard orientation:" in nline[i]:
+            return True
+
+    return False
+
+
+def get_coord_gaussian(f):
+    """Extract XYZ coordinate from Gaussian output file
+
+    :param f: user input filename
+    :type f: str
+    :return a_full: full atomic labels of complex
+    :return c_full: full atomic coordinates of complex
+    :rtype a_full: list
+    :rtype c_full: array
+    """
+    gaussian_file = open(f, "r")
+    nline = gaussian_file.readlines()
+
+    start = 0
+    end = 0
+    a_full, c_full = [], []
+    for i in range(len(nline)):
+        if "Standard orientation:" in nline[i]:
+            start = i
+
+    for i in range(start + 5, len(nline)):
+        if "---" in nline[i]:
+            end = i
+            break
+
+    for line in nline[start + 5: end]:
+        data = line.split()
+        data1 = int(data[1])
+        coord_x = float(data[3])
+        coord_y = float(data[4])
+        coord_z = float(data[5])
+        data1 = elements.check_atom(data1)
+        a_full.append(data1)
+        c_full.append([coord_x, coord_y, coord_z])
+
+    gaussian_file.close()
+
+    return a_full, c_full
+
+
+def check_nwchem_file(f):
+    """Check if the input file is NWChem file format
+
+    :param f: user input filename
+    :type f: str
+    :return: if file is NWChem output file, return True
+    :rtype: bool
+    """
+    nwchem_file = open(f, "r")
+    nline = nwchem_file.readlines()
+    for i in range(len(nline)):
+        if "No. of atoms" in nline[i]:
+            if not int(nline[i].split()[4]):
+                return False
+
+    for j in range(len(nline)):
+        if "Optimization converged" in nline[j]:
+            return True
+
+    return False
+
+
+def get_coord_nwchem(f):
+    """Extract XYZ coordinate from NWChem output file
+
+    :param f: user input filename
+    :type f: str
+    :return a_full: full atomic labels of complex
+    :return c_full: full atomic coordinates of complex
+    :rtype a_full: list
+    :rtype c_full: array
+    """
+    nwchem_file = open(f, "r")
+    nline = nwchem_file.readlines()
+
+    start = 0
+    end = 0
+    a_full, c_full = [], []
+    for i in range(len(nline)):
+        if "Optimization converged" in nline[i]:
+            start = i
+
+    for i in range(len(nline)):
+        if "No. of atoms" in nline[i]:
+            end = int(nline[i].split()[4])
+
+    start = start + 18
+    end = start + end
+    # The 1st line of coordinate is at 18 lines next to 'Optimization converged'
+    for line in nline[start:end]:
+        dat = line.split()
+        dat1 = int(float(dat[2]))
+        coord_x = float(dat[3])
+        coord_y = float(dat[4])
+        coord_z = float(dat[5])
+        dat1 = elements.check_atom(dat1)
+        a_full.append(dat1)
+        c_full.append([coord_x, coord_y, coord_z])
+
+    nwchem_file.close()
+
+    return a_full, c_full
+
+
+def check_orca_file(f):
+    """Check if the input file is ORCA file format
+
+    :param f: user input filename
+    :type f: str
+    :return: if file is ORCA output file, return True
+    :rtype: bool
+    """
+    orca_file = open(f, "r")
+    nline = orca_file.readlines()
+    for i in range(len(nline)):
+        if "CARTESIAN COORDINATES (ANGSTROEM)" in nline[i]:
+            return True
+
+    return False
+
+
+def get_coord_orca(f):
+    """Extract XYZ coordinate from ORCA output file
+
+    :param f: user input filename
+    :type f: str
+    :return a_full: full atomic labels of complex
+    :return c_full: full atomic coordinates of complex
+    :rtype a_full: list
+    :rtype c_full: array
+    """
+    orca_file = open(f, "r")
+    nline = orca_file.readlines()
+
+    start = 0
+    end = 0
+    a_full, c_full = [], []
+    for i in range(len(nline)):
+        if "CARTESIAN COORDINATES (ANGSTROEM)" in nline[i]:
+            start = i
+
+    for i in range(start + 2, len(nline)):
+        if "---" in nline[i]:
+            end = i - 1
+            break
+
+    for line in nline[start + 2:end]:
+        dat = line.split()
+        dat1 = dat[0]
+        coord_x = float(dat[1])
+        coord_y = float(dat[2])
+        coord_z = float(dat[3])
+        a_full.append(dat1)
+        c_full.append([coord_x, coord_y, coord_z])
+
+    orca_file.close()
+
+    return a_full, c_full
+
+
+def check_qchem_file(f):
+    """Check if the input file is Q-Chem file format
+
+    :param f: user input filename
+    :type f: str
+    :return: if file is Q-Chem output file, return True
+    :rtype: bool
+    """
+
+    qchem_file = open(f, "r")
+    nline = qchem_file.readlines()
+    for i in range(len(nline)):
+        if "OPTIMIZATION CONVERGED" in nline[i]:
+            return True
+
+    return False
+
+
+def get_coord_qchem(f):
+    """Extract XYZ coordinate from Q-Chem output file
+
+    :param f: user input filename
+    :type f: str
+    :return a_full: full atomic labels of complex
+    :return c_full: full atomic coordinates of complex
+    :rtype a_full: list
+    :rtype c_full: array
+    """
+    orca_file = open(f, "r")
+    nline = orca_file.readlines()
+
+    start = 0
+    end = 0
+    a_full, c_full = [], []
+    for i in range(len(nline)):
+        if "OPTIMIZATION CONVERGED" in nline[i]:
+            start = i
+
+    for i in range(start + 5, len(nline)):
+        if "Z-matrix Print:" in nline[i]:
+            end = i - 1
+            break
+
+    for line in nline[start + 5:end]:
+        dat = line.split()
+        dat1 = dat[1]
+        coord_x = float(dat[2])
+        coord_y = float(dat[3])
+        coord_z = float(dat[4])
+        a_full.append(dat1)
+        c_full.append([coord_x, coord_y, coord_z])
+
+    orca_file.close()
+
+    return a_full, c_full
