+++ octadist_gui/src/draw.py
@@ -1,17 +1,28 @@
-"""
-OctaDist  Copyright (C) 2019  Rangsiman Ketkaew et al.
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-"""
-
-from octadist import plane, elements, projection, tools, popup, main
+# OctaDist  Copyright (C) 2019  Rangsiman Ketkaew et al.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+# ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
 import matplotlib.pyplot as plt
 from mpl_toolkits.mplot3d import Axes3D
 from mpl_toolkits.mplot3d.art3d import Poly3DCollection
+
+import octadist_gui.src.popup
+import octadist_gui.src.tools
+from octadist_gui import main
+from octadist_gui.src import elements
+
 
 # import tkinter as tk
 # from matplotlib.backends.backend_tkagg import (FigureCanvasTkAgg, NavigationToolbar2Tk)
@@ -22,14 +33,14 @@ def all_atom(self, acf):
     """Display 3D structure of octahedral complex with label for each atoms
 
     :param self: master frame
-    :param acf: list - atom_coord_full
-    :return:
+    :param acf: atomic labels and coordinates of full complex
+    :type acf: list
     """
     if len(acf) == 0:
-        popup.err_no_file(self)
+        octadist_gui.src.popup.err_no_file(self)
         return 1
     elif len(acf) > 1:
-        popup.err_many_files(self)
+        octadist_gui.src.popup.err_many_files(self)
         return 1
 
     fal, fcl = acf[0]
@@ -49,7 +60,7 @@ def all_atom(self, acf):
                    color=elements.check_color(n), label="{}".format(fal[i]), s=elements.check_radii(n) * 300)
 
     # Calculate distance
-    bond_list = tools.calc_bond_distance(self, fal, fcl)
+    bond_list = octadist_gui.src.tools.find_bonds(self, fal, fcl)
     atoms_pair = []
     for i in range(len(bond_list)):
         get_atoms = bond_list[i]
@@ -111,18 +122,19 @@ def all_atom_and_face(self, acf, all_face):
     """Display 3D structure of octahedral complex with label for each atoms
 
     :param self: master frame
-    :param acf: list - atom_coord_full
-    :param all_face: list - atomic labels and coordinates of 8 faces
-    :return:
+    :param acf: atomic labels and coordinates of full complex
+    :param all_face: atomic labels and coordinates of 8 faces
+    :type acf: list
+    :type all_face: list
     """
     if len(acf) == 0:
-        popup.err_no_file(self)
+        octadist_gui.src.popup.err_no_file(self)
         return 1
     elif len(acf) > 1:
-        popup.err_many_files(self)
+        octadist_gui.src.popup.err_many_files(self)
         return 1
     if len(all_face) == 0:
-        popup.err_no_calc(self)
+        octadist_gui.src.popup.err_no_calc(self)
         return 1
 
     main.print_stdout(self, "Info: Display scattering plot of all atoms")
@@ -157,7 +169,7 @@ def all_atom_and_face(self, acf, all_face):
         ax.add_collection3d(Poly3DCollection(vertices_list[i], alpha=0.5, color=color_list[i]))
 
     # Calculate distance
-    bond_list = tools.calc_bond_distance(self, fal, fcl)
+    bond_list = octadist_gui.src.tools.find_bonds(self, fal, fcl)
     atoms_pair = []
     for i in range(len(bond_list)):
         get_atoms = bond_list[i]
@@ -198,18 +210,18 @@ def all_atom_and_face(self, acf, all_face):
     plt.show()
 
 
-def octahedron(self, aco):
+def octa(self, aco):
     """Display 3D structure of octahedral complex
 
     :param self: master frame
-    :param aco: list - atomic labels and coordinates of octahedral structures
-    :return:
+    :param aco: atomic labels and coordinates of octahedral structure
+    :type aco: list
     """
     if len(aco) == 0:
-        popup.err_no_calc(self)
+        octadist_gui.src.popup.err_no_calc(self)
         return 1
     elif len(aco) > 1:
-        popup.err_many_files(self)
+        octadist_gui.src.popup.err_many_files(self)
         return 1
 
     main.print_stdout(self, "Info: Display scattering plot of truncated octahedral structure")
@@ -267,17 +279,18 @@ def octa_and_face(self, aco, all_face):
     """Display 3D structure of octahedral complex with 8 faces
 
     :param self: master frame
-    :param aco: list - atomic labels and coordinates of octahedral structures
-    :param all_face: list - atomic labels and coordinates of 8 faces
-    :return:
+    :param aco: atomic labels and coordinates of octahedral structure
+    :param all_face: atomic labels and coordinates of 8 faces
+    :type aco: list
+    :type all_face: list
     """
     if len(aco) == 0:
-        popup.err_no_calc(self)
+        octadist_gui.src.popup.err_no_calc(self)
         return 1
     elif len(aco) > 1:
-        popup.err_many_files(self)
+        octadist_gui.src.popup.err_many_files(self)
         return 1
-    
+
     main.print_stdout(self, "Info: Display scattering plot of truncated octahedral structure")
     main.print_stdout(self, "      Draw surface for all 8 faces of selected octahedral structure")
     main.print_stdout(self, "")
@@ -342,179 +355,4 @@ def octa_and_face(self, aco, all_face):
 
     # plt.axis('equal')
     # plt.axis('off')
-    plt.show()
-
-
-def proj_planes(self, aco, all_face):
-    """Display the selected 4 faces of octahedral complex
-
-    :param self: master frame
-    :param aco: list - atomic labels and coordinates of octahedral structures
-    :param all_face: list - atomic labels and coordinates of 8 faces
-    :return:
-    """
-    if len(aco) == 0:
-        popup.err_no_calc(self)
-        return 1
-    elif len(aco) > 1:
-        popup.err_many_files(self)
-        return 1
-
-    main.print_stdout(self, "Info: Display the selected 4 pairs of opposite planes (faces)")
-    main.print_stdout(self, "      Scattering plot of all atoms")
-    main.print_stdout(self, "      Draw surface for 4 pairs of reference and opposite planes")
-    main.print_stdout(self, "")
-
-    num, metal, ao, co = aco[0]
-    # num = number of file, metal = metal center
-    # ao = atomic labels, co = atomic coordinates
-
-    face_data = all_face[0]
-    a_ref, c_ref, a_oppo, c_oppo = face_data
-
-    # reference face
-    ref_vertices_list = []
-    for i in range(4):
-        get_vertices = c_ref[i].tolist()
-        x, y, z = zip(*get_vertices)
-        vertices = [list(zip(x, y, z))]
-        ref_vertices_list.append(vertices)
-
-    # opposite face
-    oppo_vertices_list = []
-    for i in range(4):
-        x, y, z = zip(*c_oppo[i])
-        vertices = [list(zip(x, y, z))]
-        oppo_vertices_list.append(vertices)
-
-    fig = plt.figure()
-    st = fig.suptitle("4 pairs of opposite planes", fontsize="x-large")
-
-    # Display four planes
-    color_list_1 = ["red", "blue", "orange", "magenta"]
-    color_list_2 = ["green", "yellow", "cyan", "brown"]
-    for i in range(4):
-        ax = fig.add_subplot(2, 2, int(i+1), projection='3d')
-        ax.set_title("Pair {}".format(i+1))
-        ax.scatter(co[0][0], co[0][1], co[0][2], color='yellow', marker='o', s=100,
-                   linewidths=1, edgecolors='black', label="Metal center")
-        ax.text(co[0][0] + 0.1, co[0][1] + 0.1, co[0][2] + 0.1, ao[0], fontsize=9)
-
-        for j in range(1, 7):
-            ax.scatter(co[j][0], co[j][1], co[j][2], color='red', marker='o', s=50,
-                       linewidths=1, edgecolors='black', label="Ligand atoms")
-            ax.text(co[j][0] + 0.1, co[j][1] + 0.1, co[j][2] + 0.1, "{0},{1}".format(ao[j], j), fontsize=9)
-
-        # Draw plane
-        ax.add_collection3d(Poly3DCollection(ref_vertices_list[i], alpha=0.5, color=color_list_1[i]))
-        ax.add_collection3d(Poly3DCollection(oppo_vertices_list[i], alpha=0.5, color=color_list_2[i]))
-
-        # Set axis
-        ax.set_xlabel(r'X', fontsize=10)
-        ax.set_ylabel(r'Y', fontsize=10)
-        ax.set_zlabel(r'Z', fontsize=10)
-        ax.grid(True)
-
-    # Shift subplots down
-    st.set_y(1.0)
-    fig.subplots_adjust(top=0.25)
-
-    # plt.axis('equal')
-    plt.tight_layout()
-    plt.show()
-
-
-def twisting_faces(self, aco, all_face):
-    """Display twisting triangular faces and vector projection
-
-    :param self: master frame
-    :param aco: list - atomic labels and coordinates of octahedral structures
-    :param all_face: list - atomic labels and coordinates of 8 faces
-    :return:
-    """
-    if len(aco) == 0:
-        popup.err_no_calc(self)
-        return 1
-    elif len(aco) > 1:
-        popup.err_many_files(self)
-        return 1
-
-    main.print_stdout(self, "Info: Display the reference and projected atoms")
-    main.print_stdout(self, "      Scattering plot of all projected atoms on the reference plane")
-    main.print_stdout(self, "      Draw surface for 4 pairs of two twisting triangular faces")
-    main.print_stdout(self, "")
-
-    num, metal, ao, co = aco[0]
-    # num = number of file, metal = metal center
-    # ao = atomic labels, co = atomic coordinates
-
-    face_data = all_face[0]
-    a_ref, c_ref, a_oppo, c_oppo = face_data
-
-    ref_vertices_list = []
-    for i in range(4):
-        get_vertices = c_ref[i].tolist()
-        x, y, z = zip(*get_vertices)
-        vertices = [list(zip(x, y, z))]
-        ref_vertices_list.append(vertices)
-
-    fig = plt.figure()
-    st = fig.suptitle("Projected twisting triangular faces", fontsize="x-large")
-
-    for i in range(4):
-        a, b, c, d = plane.find_eq_of_plane(c_ref[i][0], c_ref[i][1], c_ref[i][2])
-        m = projection.project_atom_onto_plane(co[0], a, b, c, d)
-        ax = fig.add_subplot(2, 2, int(i+1), projection='3d')
-        ax.set_title("Projection plane {0}".format(i+1), fontsize='10')
-
-        # Projected metal center atom
-        ax.scatter(m[0], m[1], m[2], color='orange', s=100, marker='o',
-                   linewidths=1, edgecolors='black', label="Metal center")
-        ax.text(m[0] + 0.1, m[1] + 0.1, m[2] + 0.1, "{0}'".format(ao[0]), fontsize=9)
-
-        # Reference atoms
-        pl = []
-        for j in range(3):
-            ax.scatter(c_ref[i][j][0], c_ref[i][j][1], c_ref[i][j][2], color='red', s=50, marker='o',
-                       linewidths=1, edgecolors='black', label="Reference atom")
-            ax.text(c_ref[i][j][0] + 0.1, c_ref[i][j][1] + 0.1, c_ref[i][j][2] + 0.1, "{0}".format(j+1), fontsize=9)
-            # Project ligand atom onto the reference face
-            pl.append(projection.project_atom_onto_plane(c_oppo[i][j], a, b, c, d))
-
-        # Projected opposite atoms
-        for j in range(3):
-            ax.scatter(pl[j][0], pl[j][1], pl[j][2], color='blue', s=50, marker='o',
-                       linewidths=1, edgecolors='black', label="Projected ligand atom")
-            ax.text(pl[j][0] + 0.1, pl[j][1] + 0.1, pl[j][2] + 0.1, "{0}'".format(j+1), fontsize=9)
-
-        # Draw plane
-        x, y, z = zip(*pl)
-        projected_oppo_vertices_list = [list(zip(x, y, z))]
-        ax.add_collection3d(Poly3DCollection(ref_vertices_list[i], alpha=0.5, color="yellow"))
-        ax.add_collection3d(Poly3DCollection(projected_oppo_vertices_list, alpha=0.5, color="blue"))
-
-        # Draw line
-        for j in range(3):
-            merge = list(zip(m.tolist(), c_ref[i][j].tolist()))
-            x, y, z = merge
-            ax.plot(x, y, z, 'k-', color="black")
-
-        for j in range(3):
-            merge = list(zip(m.tolist(), pl[j].tolist()))
-            x, y, z = merge
-            ax.plot(x, y, z, 'k->', color="black")
-
-        # Set axis
-        ax.set_xlabel(r'X', fontsize=10)
-        ax.set_ylabel(r'Y', fontsize=10)
-        ax.set_zlabel(r'Z', fontsize=10)
-        ax.grid(True)
-
-    # Shift subplots down
-    st.set_y(1.0)
-    fig.subplots_adjust(top=0.25)
-
-    # plt.legend(bbox_to_anchor=(1.05, 1), loc=2)
-    # plt.axis('equal')
-    plt.tight_layout()
     plt.show()
