+++ octadist_gui/src/elements.py
@@ -1,11 +1,18 @@
-"""
-OctaDist  Copyright (C) 2019  Rangsiman Ketkaew et al.
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-"""
+# OctaDist  Copyright (C) 2019  Rangsiman Ketkaew et al.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+# ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
 import numpy as np
 
@@ -13,7 +20,7 @@ import numpy as np
 def check_atom(x):
     """Convert atomic number to symbol and vice versa for atom 1-109
 
-    :param x: If x is atomic number, return symbol, and vice versa.
+    :param x: If x is atomic number, return symbol, and vice-versa
     :return: symbol or atomic number, depending on input x
     """
     atoms = ['0',
@@ -69,6 +76,7 @@ def check_radii(x):
 
 def check_color(x):
     """Convert atomic number to color: 1-109
+
     Ref: http://jmol.sourceforge.net/jscolors/
 
     :param x: atomic number
