+++ octadist_gui/src/util.py
@@ -1,337 +1,45 @@
-"""
-OctaDist  Copyright (C) 2019  Rangsiman Ketkaew
+# OctaDist  Copyright (C) 2019  Rangsiman Ketkaew et al.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+# ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-"""
-
-from octadist import elements, linear, popup, main
+import functools
+import tkinter as tk
+from tkinter import scrolledtext as tkscrolled
 
 import numpy as np
-import tkinter as tk
-import tkinter.scrolledtext as tkscrolled
+import rmsd
 import scipy.optimize
-import functools
-import matplotlib.pyplot as plt
+from matplotlib import pyplot as plt
 from mpl_toolkits.mplot3d import Axes3D
 
-
-def data_complex(self, fl, facl):
-    """Show info of input complex
-
-    :param self: master frame
-    :param fl: file_list - list containing the names of all input files
-    :param facl: full_atom_coord_list
-    :return:
-    """
-    if len(fl) == 0:
-        popup.err_no_file(self)
-        return 1
-
-    main.print_stdout(self, "Info: Show info of input complex")
-    main.print_stdout(self, "")
-
-    master = tk.Toplevel(self.master)
-    master.title("Complex info")
-    master.geometry("550x500")
-    master.option_add("*Font", "Arial 10")
-    frame = tk.Frame(master)
-    frame.grid()
-
-    box = tkscrolled.ScrolledText(frame, wrap="word", width="75", height="30", undo="True")
-    box.grid(row=0, pady="5", padx="5")
-    box.delete(1.0, tk.END)
-
-    for i in range(len(fl)):
-        box.insert(tk.END, "File {0:>2} : {1}\n".format(i + 1, fl[i].split('/')[-1]))
-        box.insert(tk.END, ">> Number of atoms: {0}\n".format(len(facl[i][0])))
-        atoms = list(set(facl[i][0]))
-        box.insert(tk.END, ">> List of atoms: {0}".format(atoms) + "\n\n")
-
-
-def data_face(self, sfa, sfc, sofa, sofc):
-    """Show info of selected 4 octahedral faces
-
-    :param self: master frame
-    :param sfa: selected face atom
-    :param sfc: selected face coordinates
-    :param sofa: selected opposite face atom
-    :param sofc: selected opposite face coordinates
-    :return:
-    """
-    if len(sfa) == 0:
-        popup.err_no_calc(self)
-        return 1
-
-    main.print_stdout(self, "Info: Show info of selected 4 octahedral faces")
-    main.print_stdout(self, "")
-
-    master = tk.Toplevel(self.master)
-    master.title("Selected octahedral faces")
-    master.geometry("550x500")
-    master.option_add("*Font", "Arial 10")
-    frame = tk.Frame(master)
-    frame.grid()
-    box = tkscrolled.ScrolledText(frame, wrap="word", width="75", height="30", undo="True")
-    box.grid(row=0, pady="5", padx="5")
-    box.delete(1.0, tk.END)
-
-    for i in range(4):
-        box.insert(tk.END, "Reference atoms: {0}          Opposite atoms: {1}\n".format(sfa[i], sofa[i]))
-        for j in range(3):
-            box.insert(tk.END, "{0:9.6f},{1:9.6f},{2:9.6f} \t {3:9.6f},{4:9.6f},{5:9.6f}\n"
-                       .format(sfc[i][j][0], sfc[i][j][1], sfc[i][j][2], sofc[i][j][0], sofc[i][j][1], sofc[i][j][2]))
-        box.insert(tk.END, "\n\n")
-
-
-def param_complex(self, acf):
-    """Show structural parameters of the complex
-
-    :param self: master frame
-    :param acf: list - atom_coord_full
-    :return:
-    """
-    if len(acf) == 0:
-        popup.err_no_file(self)
-        return 1
-    elif len(acf) > 1:
-        popup.err_many_files(self)
-        return 1
-
-    main.print_stdout(self, "Info: Show structural parameters of the complex")
-    main.print_stdout(self, "")
-
-    master = tk.Toplevel(self.master)
-    master.title("Results")
-    master.geometry("380x530")
-    master.option_add("*Font", "Arial 10")
-    frame = tk.Frame(master)
-    frame.grid()
-    lbl = tk.Label(frame, text="Structural parameters of octahedral structure")
-    lbl.grid(row=0, pady="5", padx="5", sticky=tk.W)
-    box = tkscrolled.ScrolledText(frame, wrap="word", width="50", height="30", undo="True")
-    box.grid(row=1, pady="5", padx="5")
-
-    fal, fcl = acf[0]
-
-    box.insert(tk.INSERT, "Bond distance (Å)")
-    for i in range(len(fcl)):
-        for j in range(i+1, len(fcl)):
-            if i == 0:
-                distance = linear.distance_between(fcl[i], fcl[j])
-                texts = "{0}-{1}{2} {3:10.6f}".format(fal[i], fal[j], j, distance)
-            else:
-                distance = linear.distance_between(fcl[i], fcl[j])
-                texts = "{0}{1}-{2}{3} {4:10.6f}".format(fal[i], i, fal[j], j, distance)
-            box.insert(tk.END, "\n" + texts)
-
-    box.insert(tk.END, "\n\nBond angle (°)")
-    for i in range(len(fcl)):
-        for j in range(i+1, len(fcl)):
-            for k in range(j+1, len(fcl)):
-                if i == 0:
-                    angle = linear.angle_btw_3vec(self, fcl[j], fcl[i], fcl[k])
-                    texts = "{0}{1}-{2}-{3}{4} {5:10.6f}".format(fal[k], k, fal[i], fal[j], j, angle)
-                else:
-                    angle = linear.angle_btw_3vec(self, fcl[j], fcl[i], fcl[k])
-                    texts = "{0}{1}-{2}{3}-{4}{5} {6:10.6f}".format(fal[k], k, fal[i], i, fal[j], j, angle)
-                box.insert(tk.END, "\n" + texts)
-    box.insert(tk.END, "\n")
-
-
-def param_octa(self, atom_coord_octa):
-    """Show structural parameters of selected octahedral structure
-
-    :param self: master frame
-    :param atom_coord_octa: list of label and coordinate of octahedral structure
-    :return:
-    """
-    aco = atom_coord_octa
-
-    if len(aco) == 0:
-        popup.err_no_file(self)
-        return 1
-
-    main.print_stdout(self, "Info: Show structural parameters of selected octahedral structure")
-    main.print_stdout(self, "")
-
-    master = tk.Toplevel(self.master)
-    master.title("Results")
-    master.geometry("380x530")
-    master.option_add("*Font", "Arial 10")
-    frame = tk.Frame(master)
-    frame.grid()
-    lbl = tk.Label(frame, text="Structural parameters of octahedral structure")
-    lbl.grid(row=0, pady="5", padx="5", sticky=tk.W)
-    box = tkscrolled.ScrolledText(frame, wrap="word", width="50", height="30", undo="True")
-    box.grid(row=1, pady="5", padx="5")
-
-    for n in range(len(aco)):
-        if n > 0:   # separator between files
-            box.insert(tk.END, "\n\n=================================\n\n")
-
-        box.insert(tk.INSERT, "File : {0}".format(aco[n][0]))
-        box.insert(tk.END, "\nMetal: {0}".format(aco[n][1]))
-        box.insert(tk.END, "\nBond distance (Å)")
-        for i in range(7):
-            for j in range(i+1, 7):
-                if i == 0:
-                    distance = linear.distance_between(aco[n][3][i], aco[n][3][j])
-                    texts = "{0}-{1}{2} {3:10.6f}".format(aco[n][2][i], aco[n][2][j], j, distance)
-                else:
-                    distance = linear.distance_between(aco[n][3][i], aco[n][3][j])
-                    texts = "{0}{1}-{2}{3} {4:10.6f}".format(aco[n][2][i], i, aco[n][2][j], j, distance)
-                box.insert(tk.END, "\n" + texts)
-
-        box.insert(tk.END, "\n\nBond angle (°)")
-        for i in range(7):
-            for j in range(i+1, 7):
-                for k in range(j+1, 7):
-                    if i == 0:
-                        angle = linear.angle_btw_3vec(self, aco[n][3][j], aco[n][3][i], aco[n][3][k])
-                        texts = "{0}{1}-{2}-{3}{4} {5:10.6f}".format(aco[n][2][k], k, aco[n][2][i], aco[n][2][j], j, angle)
-                    else:
-                        angle = linear.angle_btw_3vec(self, aco[n][3][j], aco[n][3][i], aco[n][3][k])
-                        texts = "{0}{1}-{2}{3}-{4}{5} {6:10.6f}".format(aco[n][2][k], k, aco[n][2][i], i, aco[n][2][j], j, angle)
-                    box.insert(tk.END, "\n" + texts)
-        box.insert(tk.END, "\n")
-
-
-def calc_surface_area(self, all_face):
-    """Calculate the area of triangular face
-
-    :param self: master frame
-    :param all_face: list - atomic labels and coordinates of 8 faces
-    :return:
-    """
-    if len(all_face) == 0:
-        popup.err_no_calc(self)
-        return 1
-
-    main.print_stdout(self, "Info: Show the area of triangular face of octahedron")
-    main.print_stdout(self, "")
-
-    master = tk.Toplevel(self.master)
-    master.title("The area of triangular face")
-    master.geometry("380x500")
-    master.option_add("*Font", "Arial 10")
-    frame = tk.Frame(master)
-    frame.grid()
-    box = tkscrolled.ScrolledText(frame, wrap="word", width="50", height="30", undo="True")
-    box.grid(row=0, pady="5", padx="5")
-
-    for n in range(len(all_face)):
-        if n > 0:
-            box.insert(tk.END, "\n==============================\n\n")
-
-        face_data = all_face[n]
-        a_ref, c_ref, a_oppo, c_oppo = face_data
-
-        box.insert(tk.INSERT, "Octahedral structure no. {0}\n".format(n + 1))
-        box.insert(tk.END, "                 Atoms*        Area (Å³)\n")
-        for i in range(8):
-            area = linear.triangle_area(c_ref[i][0], c_ref[i][1], c_ref[i][2])
-            box.insert(tk.END, "Face no. {0}:  {1}      {2:10.6f}\n".format(i+1, a_ref[i], area))
-
-        box.insert(tk.END, "\n*Three ligand atoms are vertices of triangular face.\n")
-
-
-def plot_sigma_theta(self, s, t):
-    """Plot graph between Sigma and Theta parameters for multiple files
-
-    :param self: master frame
-    :param s: array - computed Sigma parameter
-    :param t: array - computed Theta parameter
-    :return:
-    """
-    if len(s) == 0:
-        popup.err_no_calc(self)
-        return 1
-
-    main.print_stdout(self, "Info: Show relationship plot between Σ and Θ")
-    main.print_stdout(self, "")
-
-    ax = plt.subplot()
-    for i in range(len(s)):
-        ax.scatter(s, t, label='Complex {0}'.format(i+1))
-        ax.text(s[i] + 0.2, t[i] + 0.2, i+1, fontsize=9)
-
-    # Shrink current axis's height by 10% on the bottom
-    box = ax.get_position()
-    ax.set_position([box.x0, box.y0 + box.height * 0.1, box.width, box.height * 0.9])
-
-    # Put a legend below current axis
-    ax.legend(loc='upper center', bbox_to_anchor=(0.5, -0.1), fancybox=True, shadow=True, ncol=5)
-
-    plt.title("Relationship plot between $\Sigma$ and $\Theta$")
-    plt.xlabel(r'$\Sigma$')
-    plt.ylabel(r'$\Theta$')
-    plt.show()
-
-
-def calc_bond_distance(self, fal, fcl):
-    """Show bond distance of atoms pair
-    - Compute distance of all bonds
-    - Screen bonds out based on global cutoff distance
-    - Screen H bonds out based on local cutoff distance
-
-    :param self: master frame
-    :param fal: list - full_atom_list
-    :param fcl: array - full_coord_list
-    :return check_2_bond_list: list - selected bonds
-    """
-    global_distance_cutoff = 2.0
-    hydrogen_distance_cutoff = 1.2
-
-    pair_list = []
-    bond_list = []
-    for i in range(len(fcl)):
-        for j in range(i+1, len(fcl)):
-            if i == 0:
-                distance = linear.distance_between(fcl[i], fcl[j])
-            else:
-                distance = linear.distance_between(fcl[i], fcl[j])
-
-            pair_list.append([fal[i], fal[j]])
-            bond_list.append([fcl[i], fcl[j], distance])
-
-    check_1_bond_list = []
-    screen_1_pair_list = []
-    for i in range(len(bond_list)):
-        if bond_list[i][2] <= global_distance_cutoff:
-            check_1_bond_list.append([bond_list[i][0], bond_list[i][1], bond_list[i][2]])
-            screen_1_pair_list.append([pair_list[i][0], pair_list[i][1]])
-
-    check_2_bond_list = []
-    for i in range(len(check_1_bond_list)):
-        if screen_1_pair_list[i][0] == "H" or screen_1_pair_list[i][1] == "H":
-            if check_1_bond_list[i][2] <= hydrogen_distance_cutoff:
-                check_2_bond_list.append([check_1_bond_list[i][0], check_1_bond_list[i][1]])
-        else:
-            check_2_bond_list.append([check_1_bond_list[i][0], check_1_bond_list[i][1]])
-
-    main.print_stdout(self, "Info: Determine the correct bond for atoms pair")
-    main.print_stdout(self, "Info: Global distance cutoff       : {0} Angstrom".format(global_distance_cutoff))
-    main.print_stdout(self, "Info: Distance cutoff for Hydrogen : {0} Angstrom".format(hydrogen_distance_cutoff))
-    main.print_stdout(self, "")
-    main.print_stdout(self, "      Total number of bonds before screening    : {0:5d}".format(len(bond_list)))
-    main.print_stdout(self, "      Total number of bonds after 1st screening : {0:5d}".format(len(check_1_bond_list)))
-    main.print_stdout(self, "      Total number of bonds after 2nd screening : {0:5d}".format(len(check_2_bond_list)))
-    main.print_stdout(self, "")
-
-    return check_2_bond_list
+import octadist_gui.src.popup
+import octadist_gui.src.tools
+from octadist_gui import main
+from octadist_gui.src import linear, elements
 
 
 def insert_text(self, text, coord, group):
-    """Insert text in box A & B
+    """Insert text in boxes
 
     :param self: root
-    :param text: string text
+    :param text: text
     :param coord: coordinates
     :param group: group A or B
-    :return:
+    :type text: str
+    :type coord: list, array
+    :type group: str
     """
     if group == "A":
         self.box_1.insert(tk.INSERT, text + "\n")
@@ -350,7 +58,6 @@ def clear_text(self):
     """Clear text in box A & B
 
     :param self: root
-    :return:
     """
     self.box_1.delete(1.0, tk.END)
     self.box_2.delete(1.0, tk.END)
@@ -371,8 +78,11 @@ def calc_fit_plane(coord):
     scipy.optimize.minimize is used to find the least-square plane
 
     :param coord: coordinates of selected atom chunk
-    :return: the equation of the plane
+    :type coord: list, array
+    :return xx, yy, z, abcd: the equation of the plane
+    :rtype xx, yy, z, abcd: list, array, tuple
     """
+
     def plane(x, y, params):
         a = params[0]
         b = params[1]
@@ -427,12 +137,21 @@ def calc_fit_plane(coord):
 
 
 def plot_fit_plane(self, acf):
-    """Plot complex and two fit planes of two sets of ligand in molecule
+    """Display complex and two fit planes of two sets of ligand in molecule
 
     :param self: root
-    :param acf: atom_coord_full
-    :return:
+    :param acf: atomic labels and coordinates of full complex
+    :type acf: list
     """
+    ###############
+    # Clear boxes #
+    ###############
+    self.box_angle1.delete(0, tk.END)
+    self.box_angle2.delete(0, tk.END)
+
+    self.box_eq1.delete(0, tk.END)
+    self.box_eq2.delete(0, tk.END)
+
     ########################
     # Find eq of the plane #
     ########################
@@ -507,7 +226,7 @@ def plot_fit_plane(self, acf):
                    color=elements.check_color(n), label="{}".format(fal[i]), s=elements.check_radii(n) * 300)
 
     # Calculate distance
-    bond_list = calc_bond_distance(self, fal, fcl)
+    bond_list = octadist_gui.src.tools.find_bonds(self, fal, fcl)
     atoms_pair = []
     for i in range(len(bond_list)):
         get_atoms = bond_list[i]
@@ -564,9 +283,10 @@ def pick_atom(self, acf, group):
     """On-mouse pick atom and get XYZ coordinate
 
     :param self: root
-    :param acf: atom_coord_full
+    :param acf: atomic labels and coordinates of full complex
     :param group: group A or B
-    :return:
+    :type acf: list
+    :type group: str
     """
     fal, fcl = acf[0]
 
@@ -586,7 +306,7 @@ def pick_atom(self, acf, group):
                    color=elements.check_color(n), label="{}".format(fal[i]), s=elements.check_radii(n) * 300)
 
     # Calculate distance
-    bond_list = calc_bond_distance(self, fal, fcl)
+    bond_list = octadist_gui.src.tools.find_bonds(self, fal, fcl)
     atoms_pair = []
     for i in range(len(bond_list)):
         get_atoms = bond_list[i]
@@ -622,12 +342,12 @@ def pick_atom(self, acf, group):
     ax.set_title('Full complex', fontsize="12")
     ax.grid(True)
 
-    # def onpick(event):
+    # def on_pick(event):
     #     ind = event.ind
-    #     print("onpick scatter:", ind, np.take(x, ind), np.take(y, ind))
+    #     print("on_pick scatter:", ind, np.take(x, ind), np.take(y, ind))
 
     # Pick point and get XYZ data
-    def onpick(event):
+    def on_pick(event):
         ind = event.ind[0]
         x, y, z = event.artist._offsets3d
         for i in range(len(fal)):
@@ -645,7 +365,7 @@ def pick_atom(self, acf, group):
                    color='yellow', s=elements.check_radii(index) * 400)
         # print(i+1, atom, x[ind], y[ind], z[ind])
 
-    fig.canvas.mpl_connect('pick_event', onpick)
+    fig.canvas.mpl_connect('pick_event', on_pick)
 
     # plt.axis('equal')
     # plt.axis('off')
@@ -656,8 +376,8 @@ def calc_jahn_teller(self, acf):
     """Calculate angular Jahn-Teller distortion parameter
 
     :param self: master
-    :param acf: atom_coord_full
-    :return:
+    :param acf: atomic labels and coordinates of full complex
+    :type acf: list
     """
     main.print_stdout(self, "Info: Calculate angular Jahn-Teller distortion parameter")
     main.print_stdout(self, "      Find least squares planes of two sets of ligands in molecule")
@@ -667,10 +387,10 @@ def calc_jahn_teller(self, acf):
     self.acf = acf
 
     if len(self.acf) == 0:
-        popup.err_no_file(self)
+        octadist_gui.src.popup.err_no_file(self)
         return 1
     elif len(self.acf) > 1:
-        popup.err_many_files(self)
+        octadist_gui.src.popup.err_many_files(self)
         return 1
 
     root = tk.Toplevel(self.master)
@@ -738,3 +458,21 @@ def calc_jahn_teller(self, acf):
 
     root.mainloop()
 
+
+def calc_rmsd(self, acf):
+    """Calculate root mean squared displacement of atoms in complex, RMSD
+
+    :param self: master
+    :param acf: atomic labels and coordinates of full complex
+    :type acf: list
+    """
+    main.print_stdout(self, "Info: Calculate root mean squared displacement of atoms in complex, RMSD")
+    main.print_stdout(self, "")
+
+    self.acf = acf
+
+    if len(self.acf) != 2:
+        octadist_gui.src.popup.err_only_2_files(self)
+        return 1
+
+    print(rmsd.__version__)
